переделал интеграцию для гугл пэй внешнего
