+++ README.md
@@ -13,7 +13,10 @@
 ## 1.4 Подключение нативного GooglePay
 
 
-## 1.5 Рекомендация в случае интеграции в flutter
+## 1.5 Подключение API внешнего взаимодействия с GooglePay
+
+
+## 1.6 Рекомендация в случае интеграции в flutter
 
 
 
@@ -216,22 +219,32 @@ dependencies {
 ## 1.4 Подключение нативного GooglePay
 
 
-* Изменить параметр isGooglePayNative в initOnCreate на true
+1) Изменить параметр isGooglePayNative в initOnCreate на true
 
-* Перейти в консоль GooglePay  https://pay.google.com/business/console/ и перейти в пункт меню Google Pay API
+2) Перейти в консоль GooglePay  https://pay.google.com/business/console/ и перейти в пункт меню Google Pay API
 
-* Найти в списке “Integrate with your Android app” ваше приложение и кликнуть “Управление”
+3) Найти в списке “Integrate with your Android app” ваше приложение и кликнуть “Управление”
 
-* Выберите тип интеграции “Через шлюз”.
+4) Выберите тип интеграции “Через шлюз”.
 
-* Загрузите скриншоты по списку указанному ниже и нажмите “Сохранить”
+5) Загрузите скриншоты по списку указанному ниже и нажмите “Сохранить”
 
-* Кликните “Submit for approval”
+6) Кликните “Submit for approval”
 
-* В случае отказа по каким-либо причинам (к примеру, дизайн кнопки не соответствует их требованиям), нужно будет ответить на письмо гугла, что вы используете стороннее решение компании Airba Pay
+7) В случае отказа по каким-либо причинам (к примеру, дизайн кнопки не соответствует их требованиям), нужно будет ответить на письмо гугла, что вы используете стороннее решение компании Airba Pay
 
 
-## 1.5 Рекомендация в случае интеграции в flutter
+## 1.5 Подключение API внешнего взаимодействия с GooglePay
+
+1) Добавьте кнопку GooglePay и код для взяимодействия с ней, используя документацию GooglePay или samples из github гугла.
+
+2) Выполните AirbaPaySdk.initOnCreate(). Если есть необходимость скрывать на страницах сдк кнопку GooglePay, то во всех случаях вызова initOnCreate используйте флаг hideInternalGooglePayButton = true. 
+
+3) 
+
+
+
+## 1.6 Рекомендация в случае интеграции в flutter
 
 Нужно создать дополнительный промежуточный активити FlutterAirbaPayActivity,
 в котором будет производиться инициализация и переход на страницу сдк.
