+++ apay_android/build.gradle
@@ -192,7 +192,7 @@ dependencies {
 
 
 group 'kz.airbapay' // this is your groupId
-version '1.0.33' // this is your version
+version '1.0.36' //_2023_08_00  // todo меняй и в BaseInterceptor в requestBuilder.addHeader("Platform", "Android,
 
 ext { // this are variables
     pomGroupId = group
