+++ apay_android/src/main/java/kz/airbapay/apay_android/ui/pages/test_page/TestActivity.kt
@@ -168,7 +168,7 @@ class TestActivity : ComponentActivity() {
                             startActivity(
                                 Intent(
                                     this@TestActivity,
-                                    TestGooglePayExternalActivity::class.java
+                                    TestComposeGooglePayExternalActivity::class.java
                                 )
                             )
                         }
